+++ src/main/java/com/example/app/service/UserService.java
@@ -2,13 +2,12 @@ package com.example.app.service;
 
 import com.example.app.domain.dto.UserDTO;
 import com.example.app.domain.dto.UserDetailsWithRole;
-import com.example.app.domain.model.Role;
 import com.example.app.domain.model.User;
 import com.example.app.repository.UserRepository;
+import com.example.app.utils.MessageException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,9 +19,16 @@ import java.util.Optional;
 public class UserService {
     private final UserRepository repository;
 
-
     @Transactional
     public User createUser(UserDTO userDTO) {
+        if (repository.existsByUsername(userDTO.getUsername())) {
+            throw new MessageException("Пользователь с таким именем уже существует");
+        }
+
+        if (repository.existsByEmail(userDTO.getEmail())) {
+            throw new MessageException("Пользователь с таким email уже существует");
+        }
+
         User user = User.builder()
                 .username(userDTO.getUsername())
                 .password(userDTO.getPassword())
@@ -30,6 +36,9 @@ public class UserService {
                 .role(userDTO.getRole())
                 .build();
         return repository.save(user);
+    }
+    public User findByUsername(String username) {
+        return repository.findByUsername(username);
     }
     public User updateUser(Long id, UserDTO userDTO) {
         return repository.findById(id)
@@ -41,11 +50,14 @@ public class UserService {
                     user.setPosition(userDTO.getPosition());
                     return repository.save(user);
                 })
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new MessageException("Пользователь с идентификатором не найден: " + id));
     }
 
     @Transactional
     public void deleteUser(Long userId) {
+        if (!repository.existsById(userId)) {
+            throw new MessageException("Пользователь с идентификатором не найден: " + userId);
+        }
         repository.deleteById(userId);
     }
 
@@ -58,82 +70,45 @@ public class UserService {
     public List<User> searchUsersByUsername(String username) {
         return repository.findByUsernameContainingIgnoreCase(username);
     }
-    /**
-     * Сохранение пользователя
-     *
-     * @return сохраненный пользователь
-     */
+
     public User save(User user) {
-        return repository.save(user);
-    }
-
-
-    /**
-     * Создание пользователя
-     *
-     * @return созданный пользователь
-     */
-    public User create(User user) {
         if (repository.existsByUsername(user.getUsername())) {
-            // Заменить на свои исключения
-            throw new RuntimeException("Пользователь с таким именем уже существует");
+            throw new MessageException("Пользователь с таким именем уже существует");
         }
 
         if (repository.existsByEmail(user.getEmail())) {
-            throw new RuntimeException("Пользователь с таким email уже существует");
+            throw new MessageException("Пользователь с таким email уже существует");
+        }
+
+        return repository.save(user);
+    }
+
+    public User create(User user) {
+        if (repository.existsByUsername(user.getUsername())) {
+            throw new MessageException("Пользователь с таким именем уже существует");
+        }
+
+        if (repository.existsByEmail(user.getEmail())) {
+            throw new MessageException("Пользователь с таким email уже существует");
         }
 
         return save(user);
     }
 
-    /**
-     * Получение пользователя по имени пользователя
-     *
-     * @return пользователь
-     */
     public User getByUsername(String username) {
-        return repository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
-
+        return repository.findByUsername(username);
     }
-
     public Optional<User> getUserById(Long id) {
         return repository.findById(id);
     }
 
-    /**
-     * Получение пользователя по имени пользователя
-     * <p>
-     * Нужен для Spring Security
-     *
-     * @return пользователь
-     */
     public UserDetailsService userDetailsService() {
         return this::getByUsername;
     }
 
-    /**
-     * Получение текущего пользователя
-     *
-     * @return текущий пользователь
-     */
     public User getCurrentUser() {
-        // Получение имени пользователя из контекста Spring Security
         var username = SecurityContextHolder.getContext().getAuthentication().getName();
         return getByUsername(username);
-    }
-
-
-    /**
-     * Выдача прав администратора текущему пользователю
-     * <p>
-     * Нужен для демонстрации
-     */
-    @Deprecated
-    public void getAdmin() {
-        var user = getCurrentUser();
-        user.setRole(Role.ROLE_ADMIN);
-        save(user);
     }
 
     public UserDetailsWithRole getCurrentUserWithRole() {
